chore(deps): taskypi@1.6.0
add taskypi as a general task runner. The `scripts` in the
pyproject.toml file are added as runnable script to the user path when
they install the package. This is not the desirable outcome and can lead
to confusing behviors.

This allows up to implement develomnet tasks similar to npm script with
out effecting the installable package or host environment