fix(threading): account for secondary buffer flush and deadlock
In the case that work cannot be scheduled on the worker pool, logs are
pushed onto the `secondary` list and the check before a send doesn't
account for any logs that may be pending in the `secondary`.

This can happen in the case of a run time error when attempting to
submit work to one of the thread pools effectively leaving logs unsent.

Additionally eliminates a situation where lock wasn't being released
during and error situation and would evenutally result in every tread
being locked. This would prevent logs from being sent

fixes: #74